+++ PhpSerializerNET/Extensions/ArrayExtensions.cs
@@ -10,6 +10,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 
 namespace PhpSerializerNET
@@ -19,6 +20,27 @@ namespace PhpSerializerNET
 		public static string Utf8Substring(this byte[] array, int start, int length)
 		{
 			return Encoding.UTF8.GetString(array.Skip(start).Take(length).ToArray());
+		}
+
+		public static PropertyInfo FindProperty(this PropertyInfo[] array, string name, PhpDeserializationOptions options){
+			PropertyInfo property = null;
+
+			// Explicetly named properties have priority:
+			property = array
+				.Where(y => y.GetCustomAttribute<PhpPropertyAttribute>() != null)
+				.FirstOrDefault(y =>
+					options.CaseSensitiveProperties
+						? y.GetCustomAttribute<PhpPropertyAttribute>().Name  == name
+						: y.GetCustomAttribute<PhpPropertyAttribute>().Name.ToLower() == name.ToLower()
+				);
+			
+			if (property == null) {
+				property = options.CaseSensitiveProperties 
+					? array.FirstOrDefault(y => y.Name == name)
+					: array.FirstOrDefault(y => y.Name.ToLower() == name.ToLower());
+			}
+
+			return property;
 		}
 	}
 }