+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+- Added option to convert strings "1" and "0" into bools when deserializing an object.
+- Changed how validation is handled and moved validation out of the tokenization step.
+- Added `[PhpIgnore]` and `[PhpProperty("name")]`
+
 # 0.1.0 
 
 - Initial release.